testsuite: split subprocess unit test
Problem: the subprocess unit test is very long and hard to follow.

Break out stdio and channel tests into new unit test programs.
